requirements.txt
Django==5.0.2
pymongo==4.6.1
django-mongodb-engine==0.6.0
python-dotenv==1.0.1
cryptography==42.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-magic==0.4.27
django-storages==1.14.2
boto3==1.34.34 